Outfactor utility functions from views
